huge refactor and change of approach

+++ Genelife.Domain/Commands/Cheat/SetHunger.cs
@@ -2,4 +2,4 @@ using MassTransit;
 
 namespace Genelife.Domain.Commands.Cheat;
 
-public record SetHunger(Guid CorrelationId, int Value) : CorrelatedBy<Guid>;
+public record SetHunger(Guid CorrelationId, float Value) : CorrelatedBy<Guid>;
